Fixed some async access functions in TibiaWiki cog - Moved legacy sqlite code to database_migration - Fixed travis job
